projects/tools/prompt-engineer/requirements-async.txt
# Additional dependencies for async performance and API features

# Async web framework and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Redis for advanced caching
redis>=5.0.0
hiredis>=2.2.0  # Faster Redis parsing

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Performance monitoring
prometheus-client>=0.18.0
psutil>=5.9.0

# Export functionality
reportlab>=4.0.0  # PDF generation
python-docx>=0.8.11  # DOCX export
jinja2>=3.1.0  # HTML templates
weasyprint>=60.0  # HTML to PDF

# Enhanced file handling
aiofiles>=23.0.0
watchdog>=3.0.0  # File system monitoring

# Background tasks and queuing
celery[redis]>=5.3.0
flower>=2.0.0  # Celery monitoring

# Development and testing
pytest-asyncio>=0.21.0
httpx>=0.25.0  # Async HTTP client for testing